Move git_version_info to util
